% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_origin.R
\name{find_origin}
\alias{find_origin}
\alias{scr_lda}
\alias{scr_nb}
\alias{scr_ml}
\title{Core functions to asses overall performance of LDA, NB and ML.}
\usage{
find_origin(
  method = c("lda", "nb", "ml"),
  ndistr = 20,
  nsample = 10,
  noise = 0,
  col_ids = 3:19,
  df_dat = NULL,
  toprob = TRUE
)

scr_lda(df_dat, ndistr = 20, nsample = 10, noise = 0, col_ids = 3:19)

scr_nb(df_dat, ndistr = 20, nsample = 5, noise = 0, col_ids = 3:19)

scr_ml(
  df_dat = NULL,
  ndistr = 20,
  nsample = 5,
  noise = 0,
  col_ids = 3:19,
  toprob = TRUE
)
}
\arguments{
\item{method}{statistical approach used, either "lda", "nb" or "ml".}

\item{ndistr}{number of sample used to build baseline biotracer distributions.}

\item{nsample}{number of samples.}

\item{noise}{noise to be added.}

\item{col_ids}{ids (name or integer) of columns to be used (select biotracers).}

\item{df_dat}{biotracer data sets. Note that by default, it loads \code{df_all()}
(see \code{\link[=get_data_ready]{get_data_ready()}}).}

\item{toprob}{should the output be a probability.}
}
\description{
Core functions to asses overall performance of LDA, NB and ML.
}
\details{
\code{ndistr + nsample} cannot exceed 30 with data in \code{df_all}.
}
\section{Functions}{
\itemize{
\item \code{scr_lda}: core function for LDA.

\item \code{scr_nb}: core function for the Naive Bayesian (NB) classifier.

\item \code{scr_ml}: core function for one hidden layer perceptron (ML) classifier.
}}

