% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isos_simulations.R
\name{isos_simulations}
\alias{isos_simulations}
\alias{isos_simulations_mixture2}
\alias{isos_sample}
\alias{isos_likelihood}
\title{simulations for isoscape}
\usage{
isos_simulations(ls_values, sample_size, ls_density, nrep)

isos_simulations_mixture2(
  ls_values,
  sample_size,
  ls_density,
  dens1,
  dens2,
  perc,
  nrep
)

isos_sample(sample, ls_density)

isos_likelihood(sample, density)
}
\arguments{
\item{ls_values}{List of values to be sampled.}

\item{sample_size}{size of the samples to be drawn.}

\item{ls_density}{a list of object of class \code{\link[stats:density]{stats::density()}}.}

\item{nrep}{an integer indicating the number of repetition (see \code{details} section).}

\item{dens1, dens2}{an integer that identities the first (and second) distribution to be mixed (in \code{ls_density}).}

\item{perc}{a vector of numeric indicating the percentage of dens1 included in the sample (assuming the remaining is made of dens2).}

\item{sample}{values of a sample.}

\item{density}{an object of class \code{\link[stats:density]{stats::density()}}.}
}
\value{
A matrix

A vector of probabilities.

A vector of probabilities of origin for every distribution in ls_density.

A likelihood value.
}
\description{
Functions to compute probability of origins to a given sample.
}
\details{
For every density included in \code{ls_density} the likelihood that the sample is
coming from a specific region is computed. Then likelihood values are
compared to determine the probability for \code{sample} of having been drawn
from the density (there is a reason why the ratio work). Probability for
all densities are returned and the corresponding vector is ordered following
\code{ls_density}.
}
\section{Functions}{
\itemize{
\item \code{isos_simulations_mixture2}: Analyse for one sample.

\item \code{isos_sample}: Analyse for one sample.

\item \code{isos_likelihood}: Computes the likelihood associated with one sample.
}}

