% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fig2.R
\name{scr_fig2}
\alias{scr_fig2}
\title{Figure 2}
\usage{
scr_fig2(
  path = "output/res_lda_nb/nbio/non_pca",
  file_ml = "output/res_f/res_combn_ml_nbio.rds"
)
}
\arguments{
\item{path}{Path to results files for LDA and NBC.}

\item{file_ml}{Path to results file (a rds file) for MLP.}
}
\description{
Code to reproduce figure 2 (increase bio-tracers).
}
