% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fig3.R
\name{scr_fig3}
\alias{scr_fig3}
\alias{scr_figS7}
\title{Figure 3}
\usage{
scr_fig3(
  path = "output/res_lda_nb/noise/",
  file_ml = "output/res_f/res_combn_ml_nbio.rds"
)

scr_figS7()
}
\arguments{
\item{path}{Path to results files for LDA and NBC.}

\item{file_ml}{Path to results file (a rds file) for MLP.}
}
\description{
Code to reproduce figure 3 (noise addition).
}
\section{Functions}{
\itemize{
\item \code{scr_figS7}: Code to reproduce figure S7.
}}

