% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fig4.R
\name{scr_fig4}
\alias{scr_fig4}
\title{Figure 4}
\usage{
scr_fig4(
  path = "output/res_lda_nb/nbio/non_pca",
  file_ml = "output/res_f/res_combn_ml_nbio.rds",
  files_pca = c("output/res_f/nbio_pca_lda_01.rds", "output/res_f/nbio_pca_nb_01.rds",
    "output/res_f/ml_nbio_pca.rds")
)
}
\arguments{
\item{path}{Path to results files for LDA and NBC.}

\item{file_ml}{Path to results file (a rds file) for MLP.}

\item{files_pca}{A vector of 3 rds files (1 per method) corresponding to the results obtained once a PCA was applied to the data set.}
}
\description{
Code to reproduce figure 4.
}
