% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fig_concept.R
\name{scr_fig_concept}
\alias{scr_fig_concept}
\title{Figure 1}
\usage{
scr_fig_concept()
}
\description{
Code to reproduce figure 1
}
\details{
Note that the Salmon silhouette was added with Inkscape and is available at
\url{http://phylopic.org/image/3c098bb8-4158-4777-b567-80e48049409c/}.
}
