% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_ks.R
\name{scr_lda}
\alias{scr_lda}
\alias{scr_ks}
\title{Core functions to asses overall performance of LDA and KS.}
\usage{
scr_lda(
  ndistr = 20,
  nsample = 10,
  noise = 0,
  col_ids = 3:19,
  df_dat = NULL,
  toprob = TRUE
)

scr_ks(
  ndistr = 25,
  nsample = 5,
  noise = 0,
  col_ids = c("d13C", "d15N", "d34S"),
  df_dat = NULL,
  toprob = TRUE
)
}
\arguments{
\item{ndistr}{number of sample used to build baseline biotracer distributions.}

\item{nsample}{number of}

\item{noise}{noise to be added.}

\item{col_ids}{ids (name or integer) of columns to be used (select biotracers).}

\item{df_dat}{biotracer data sets. Note that by default, it loads \code{df_all()}
(see \code{\link[=get_data_ready]{get_data_ready()}}).}

\item{toprob}{should the output be a probability.}
}
\description{
Core functions to asses overall performance of LDA and KS.
}
\details{
\code{ndistr + nsample} cannot exceed 30 with data in \code{df_all}
}
\section{Functions}{
\itemize{
\item \code{scr_ks}: core function for KS
}}

