% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_simu.R
\name{scr_simu}
\alias{scr_simu}
\alias{simu_nbio}
\alias{simu_ndistr}
\alias{simu_nsample}
\title{Simulation functions}
\usage{
scr_simu()

simu_nbio(nrep = 200, mxcb = 200, nsample, ndistr, noise)

simu_ndistr(nrep = 200, mxcb = 200, nbio = 5, nsample, noise)

simu_nsample(nrep = 500, mxcb = 700, nbio = 3, ndistr, noise)
}
\arguments{
\item{nrep}{number of repetition for every iterations.}

\item{mxcb}{maximum number of biotracers combinations (see details).}

\item{nsample}{number of samples (individuals) to be tested (per region).}

\item{ndistr}{number of individuals used to generate distributions (per region).}

\item{noise}{noise to be added.}

\item{nbio}{number of biotracers to be used.}
}
\description{
Simulation functions
}
\details{
If the total number of combinations is smaller than \code{mxcb} then the
total number o of combinations is used.
}
\section{Functions}{
\itemize{
\item \code{simu_nbio}: simulation for an increasing number of biotracers.

\item \code{simu_ndistr}: increase the number of samples used to build the distribution.

\item \code{simu_nsample}: increase the number of samples used to build the distribution.
}}

