% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_theory.R
\name{scr_theory}
\alias{scr_theory}
\title{Theory script}
\usage{
scr_theory(nrep = 10000, nrep2 = 1e+05, mx_sp_sz = 50, rerun = FALSE)
}
\arguments{
\item{nrep}{number of replicates for figure S1, S3 and S4.}

\item{nrep2}{number of replicates for figures S1, S5 and S6.}

\item{mx_sp_sz}{maximum sample size value.}

\item{rerun}{a logical. Should simulations be rerun for figures S5 and S6.}
}
\description{
Code to reproduce analysis in the Supplementary information. Also, this
script creates figures S1, S2, S3 and S4.
}
