% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fig5.R
\name{scr_fig5}
\alias{scr_fig5}
\title{Figures 5, 6 and 7}
\usage{
scr_fig5(file = "output/res_lda_nb/all123/res_lda_123.rds")
}
\arguments{
\item{file}{path to results for individual performances as well as performances of all pairs and triplets.}
}
\description{
Code to reproduce figures 5, 6 and 7.
}
