% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_theory.R
\name{scr_theory}
\alias{scr_theory}
\title{Theory script}
\usage{
scr_theory(nrep = 10000, mx_sp_sz = 50)
}
\arguments{
\item{nrep}{number of replicates.}

\item{mx_sp_sz}{maximum sample size value.}
}
\description{
Code to reproduce analysis in the Supplementary information. Also, this
script creates figures S1, S2, S3 and S4.
}
