% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_simulations.R
\name{simu_nbio}
\alias{simu_nbio}
\alias{simu_nbio_order}
\alias{simu_ndistr}
\alias{simu_nsample}
\alias{simu_noise}
\title{Simulation functions}
\usage{
simu_nbio(
  method = c("lda", "nb", "ml"),
  nrep = 100,
  mxcb = 10,
  nsample = 10,
  ndistr = 20,
  noise = 0,
  pca = FALSE,
  mx_bio = 17,
  ...
)

simu_nbio_order(
  method = c("lda", "nb", "ml"),
  nrep = 100,
  nsample = 10,
  ndistr = 20,
  noise = 0
)

simu_ndistr(
  method = c("lda", "nb", "ml"),
  nrep = 20,
  mxcb = 20,
  nbio = 5,
  nsample = 1,
  noise = 0,
  pca = FALSE
)

simu_nsample(
  method = c("lda", "nb", "ml"),
  nrep = 20,
  mxcb = 20,
  nbio = 5,
  ndistr = 20,
  noise = 0,
  pca = FALSE
)

simu_noise(
  method = c("lda", "nb", "ml"),
  nrep = 20,
  mxcb = 20,
  nbio = 5,
  ndistr = 20,
  nsample = 5,
  pca = FALSE
)
}
\arguments{
\item{method}{statistical approach, either "lda", "nb" or "ml".}

\item{nrep}{number of repetition for every iterations.}

\item{mxcb}{maximum number of biotracers combinations (see details).}

\item{nsample}{number of samples (individuals) to be tested (per region).}

\item{ndistr}{number of individuals used to generate distributions (per region).}

\item{noise}{noise to be added.}

\item{pca}{a logical. See \code{\link[=get_data_ready]{get_data_ready()}}.}

\item{mx_bio}{maximum biotracer number. '}

\item{...}{further arguments for "ml".}

\item{nbio}{number of biotracers to be used.}
}
\description{
Simulation functions
}
\details{
If the total number of combinations is smaller than \code{mxcb} then the
total number of combinations is used.
}
\section{Functions}{
\itemize{
\item \code{simu_nbio_order}: same as \code{\link[=simu_nbio]{simu_nbio()}} but use pca and keep the axes ordered.

\item \code{simu_ndistr}: effect of the number of samples used to build the distribution.

\item \code{simu_nsample}: effect of the number of samples used for inference.

\item \code{simu_noise}: effect of the number of samples used for inference.
}}

\examples{
\dontrun{
simu_nbio("lda")
simu_ndistr("lda")
simu_nsample()
}
}
