% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_simulations.R
\name{simu_nbio}
\alias{simu_nbio}
\alias{simu_ndistr}
\title{Simulation functions}
\usage{
simu_nbio(
  method = c("lda", "nb", "ml"),
  nrep = 100,
  mxcb = 10,
  nsample = 10,
  ndistr = 20,
  noise = 0
)

simu_ndistr(
  method = c("lda", "nb", "ml"),
  nrep = 20,
  mxcb = 20,
  nbio = 5,
  nsample = 1,
  noise = 0
)
}
\arguments{
\item{method}{statistical approach, either "lda", "nb" or "ml".}

\item{nrep}{number of repetition for every iterations.}

\item{mxcb}{maximum number of biotracers combinations (see details).}

\item{nsample}{number of samples (individuals) to be tested (per region).}

\item{ndistr}{number of individuals used to generate distributions (per region).}

\item{noise}{noise to be added.}

\item{nbio}{number of biotracers to be used.}
}
\description{
Simulation functions
}
\details{
If the total number of combinations is smaller than \code{mxcb} then the
total number of combinations is used.
}
\section{Functions}{
\itemize{
\item \code{simu_ndistr}: effect of the number of samples used to build the distribution.
}}

\examples{
\dontrun{
simu_nbio("lda")
simu_ndistr("lda")
}
}
